chore: update pytest
