+++ omnisafe/envs/__init__.py
@@ -15,14 +15,13 @@
 """Environment API for OmniSafe."""
 
 from omnisafe.envs import classic_control
-from omnisafe.envs.classic_control.envs_from_cbf import BarrierFunctionEnv
-from omnisafe.envs.classic_control.envs_from_rcbf import RobustBarrierFunctionEnv
+from omnisafe.envs.cbf_env import BarrierFunctionEnv
 from omnisafe.envs.core import CMDP, env_register, make, support_envs
 from omnisafe.envs.crabs_env import CRABSEnv
 from omnisafe.envs.custom_env import CustomEnv
 from omnisafe.envs.meta_drive_env import SafetyMetaDriveEnv
-from omnisafe.envs.barrier_function_env import BarrierFunctionEnv
 from omnisafe.envs.mujoco_env import MujocoEnv
+from omnisafe.envs.rcbf_env import RobustBarrierFunctionEnv
 from omnisafe.envs.safety_gymnasium_env import SafetyGymnasiumEnv
 from omnisafe.envs.safety_gymnasium_modelbased import SafetyGymnasiumModelBased
 from omnisafe.envs.safety_isaac_gym_env import SafetyIsaacGymEnv
